+++ src/Router.jsx
@@ -3,7 +3,7 @@ import { BrowserRouter as Router, Route, Switch } from 'react-router-dom'
 import { LANDING_PAGE_PATH } from '@/constants'
 
 import withSecure from '@/components/wrappers/withSecure'
-import Landing from '@/components/pages/landing'
+import Landing from '@/components/pages/Landing'
 import NotFound from '@/components/pages/NotFound'
 
 export default () => (
